+++ NEWS.md
@@ -1,5 +1,15 @@
-# flashlight 0.1.0
+# flashlight 0.2.0
 
-* Initial release
+* New functionality
+
+- Added variable contribution breakdown for single observations.
+
+* Interface change
+
+- Removed `zero_counts` argument in `plot_counts`.
+
+* Bug fixes
+
+- `zero_counts` in `plot_counts` and `plot.light_effects` had no effect for single flashlights with no "by" variable. This is fixed.